pull VideoNumberRoll out of VideoView
